Move resource.URN to urn.URN
This change follows on https://github.com/pulumi/pulumi/pull/15157, to allow
https://github.com/pulumi/pulumi/pull/15145 to be imported by `resource`. This is the last
module import cycle change necessary for `resource` to import `property` (as in
https://github.com/pulumi/pulumi/pull/15145).

Like https://github.com/pulumi/pulumi/pull/15157, this change is *100% backwards*
compatible.
